# api/openapi.yaml
openapi: 3.1.0
info:
  title: Halo management API
  version: 0.1.0
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The JSON API behind the Halo console and account portal. Anything the web interface does, a script can do with the same requests.

    **Authentication.** Send either the `halo_session` cookie of a signed-in person, or an API key of a service account as `Authorization: Bearer hlk_…`. API keys work only when they have the `api` scope, act with the roles of their service account, and are ignored on the personal routes under `/me` and on the sign-in routes under `/auth`. The personal routes under `/me/` (with a trailing segment, not `/me` itself) also accept an access token that `halo login` obtained for the built-in Halo CLI application.

    **Roles.** Each operation states the roles it requires. A global administrator passes every role check. "Any administrator" means security, user, helpdesk or application administrator, or auditor.

    **Errors.** Every error is `{"error": {"code": "ERR_…", "message": "…"}}`. Match on `code`; show `message` to people. Besides the codes listed per operation, any request can fail with `500` and `ERR_INTERNAL`, and unknown routes answer `404` with `ERR_NO_ROUTE`. Requests with a method other than GET, HEAD or OPTIONS that come from another website get `403` with `ERR_CROSS_SITE`.

...
servers:
  - url: /api/v1
security:
  - session: []
  - apiKey: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List every person
      description: 'Roles: any administrator. Service accounts are not included.'
      responses:
        '200':
          description: Every person, sorted by name.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Users]
      operationId: inviteUser
      summary: Invite a person
      description: 'Roles: user administrator. Creates the account with status `invited` and a setup link, and emails the link when SMTP is configured.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name]
              properties:
                email: {type: string, format: email, maxLength: 254}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string, description: Id of another person.}
                groupIds:
                  type: array
                  description: Assigned groups only.
                  items: {type: string}
      responses:
        '201':
          description: The new person and their setup link.
          content:
            application/json:
              schema:
                type: object
                required: [user, enrollUrl, expiresAt, emailed]
                properties:
                  user: {$ref: '#/components/schemas/User'}
                  enrollUrl: {type: string, format: uri}
                  expiresAt: {type: string, format: date-time}
                  emailed: {type: boolean, description: True when Halo also emailed the link.}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    get:
      tags: [Users]
      operationId: getUser
      summary: Get one person
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    patch:
      tags: [Users]
      operationId: updateUser
      summary: Change a person's profile
      description: 'Roles: user or helpdesk administrator. Helpdesk administrators can change title, department and location only. Only a global administrator can change a person who holds an administrator role. Send only the fields to change; an empty `managerId` removes the manager.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: {type: string, format: email}
                name: {type: string, maxLength: 200}
                title: {type: string, maxLength: 200}
                department: {type: string, maxLength: 200}
                location: {type: string, maxLength: 200}
                managerId: {type: string}
      responses:
        '200':
          description: The updated person.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Invalid'}
...
  /users/{id}/suspend:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: suspendUser
      summary: Suspend a person and end their sessions
      description: 'Roles: user or security administrator. Refused for your own account and for the last active global administrator.'
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
  /users/{id}/restore:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: restoreUser
      summary: Restore a suspended person
      description: 'Roles: user administrator. The person becomes `active`, or `invited` when they never signed in and have no sign-in method.'
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /users/{id}/revoke-sessions:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: revokeUserSessions
      summary: End a person's sessions
      description: 'Roles: user, helpdesk or security administrator. When you target yourself, your current session stays.'
      responses:
        '200': {$ref: '#/components/responses/Revoked'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /users/{id}/reset-authentication:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    post:
      tags: [Users]
      operationId: resetAuthentication
      summary: Remove every sign-in method and create a new setup link
      description: 'Roles: user or helpdesk administrator. Also ends every session and emails the link when SMTP is configured.'
      responses:
        '200':
          description: The new setup link.
          content:
            application/json:
              schema:
                type: object
                required: [enrollUrl, expiresAt, emailed]
                properties:
                  enrollUrl: {type: string, format: uri}
                  expiresAt: {type: string, format: date-time}
                  emailed: {type: boolean}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /users/{id}/roles:
    parameters:
      - {$ref: '#/components/parameters/Id'}
    put:
      tags: [Users]
      operationId: setUserRoles
      summary: Replace a person's administrator roles
      description: 'Roles: global administrator. Send the complete list; an empty list removes every role. You cannot remove your own global administrator role.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items: {$ref: '#/components/schemas/RoleKey'}
      responses:
        '200': {$ref: '#/components/responses/UserResponse'}
        '400': {$ref: '#/components/responses/BadJSON'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Invalid'}
...
  /sessions:
    get:
      tags: [Activity]
      operationId: listSessions
      summary: List every active session
      description: 'Roles: any administrator.'
      responses:
        '200':
          description: The sessions.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Session'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /sign-ins:
    get:
      tags: [Activity]
      operationId: listSignIns
      summary: List sign-in events
      description: 'Roles: any administrator.'
      parameters:
        - {name: user, in: query, description: Only this person's events., schema: {type: string}}
        - {name: app, in: query, description: Only this application's events., schema: {type: string}}
        - {$ref: '#/components/parameters/Limit'}
      responses:
        '200':
          description: The events, newest first.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/SignInEvent'}
        '401': {$ref: '#/components/responses/Unauthenticated'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '422': {$ref: '#/components/responses/Invalid'}
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: halo_session
      description: The session cookie Halo sets when a person signs in.
    apiKey:
      type: http
      scheme: bearer
      description: An API key of a service account, starting with `hlk_`, with the `api` scope. Requests act with the service account's roles.
    cliAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An access token that `halo login` obtained for the built-in Halo CLI application. Accepted only on routes under `/me/`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string}
...
    Limit:
      name: limit
      in: query
      description: How many events to return, 1 to 1000. Defaults to 500.
      schema: {type: integer, minimum: 1, maximum: 1000, default: 500}
...
  responses:
...
    BadJSON:
      description: '`ERR_INVALID_JSON`: the body is not valid JSON, is larger than 1 MiB, or has an unknown field.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthenticated:
      description: '`ERR_UNAUTHENTICATED`: no valid session or key.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Forbidden:
      description: '`ERR_FORBIDDEN`: your roles do not allow this. `ERR_CROSS_SITE` on changes from another website.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: '`ERR_NOT_FOUND`: the item does not exist.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Conflict:
      description: '`ERR_CONFLICT`: the change conflicts with the current state, such as a name that is already used.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Invalid:
      description: '`ERR_INVALID`: a field is missing or invalid. The message names it.'
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    Revoked:
      description: The number of sessions ended.
      content:
        application/json:
          schema:
            type: object
            required: [revoked]
            properties:
              revoked: {type: integer}
    UserResponse:
      description: The person.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/User'}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string, examples: [ERR_NOT_FOUND]}
            message: {type: string, examples: [This user was not found. It may have been deleted.]}
...
    MethodKind:
      type: string
      enum: [passkey, security-key, totp, magic-link, recovery-codes, federated]
    RoleKey:
      type: string
      enum: [global_admin, security_admin, user_admin, helpdesk_admin, app_admin, auditor]
...
    Method:
      type: object
      required: [id, kind, label, addedAt, lastUsedAt]
      properties:
        id: {type: string, description: '`recovery-codes` for the set of recovery codes.'}
        kind: {type: string, enum: [passkey, security-key, totp, recovery-codes]}
        label: {type: string, examples: [Passkey · Safari on iOS]}
        addedAt: {type: string, format: date-time}
        lastUsedAt: {type: [string, 'null'], format: date-time}
        detail: {type: string, examples: [8 of 10 remaining]}
    User:
      type: object
      required: [id, email, emailVerified, name, title, department, location, status, strength, methods, roles, groupIds, appIds, managerId, source, createdAt, lastSignInAt]
      properties:
        id: {type: string, examples: [usr_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
        email: {type: string, format: email}
        emailVerified: {type: boolean, description: 'True once the person redeemed a magic link, finished setup from a link Halo emailed, or signed in through an identity provider that verified the same address. Changing the email clears it. With the `email` scope, ID tokens and userinfo carry it as `email_verified`.'}
        name: {type: string}
        title: {type: string}
        department: {type: string}
        location: {type: string}
        status: {type: string, enum: [invited, active, suspended, deprovisioned]}
        strength: {type: string, enum: [phishing-resistant, multi-factor, single-factor]}
        methods:
          type: array
          items: {$ref: '#/components/schemas/Method'}
        roles:
          type: array
          description: Role names, such as `Global administrator`. Assign roles with their keys.
          items: {type: string}
        groupIds:
          type: array
          description: Assigned and rule-based groups.
          items: {type: string}
        appIds:
          type: array
          description: Applications the person is assigned to through their groups.
          items: {type: string}
        managerId: {type: [string, 'null']}
        source: {type: string, examples: [Halo, 'SCIM · Okta', Google]}
        createdAt: {type: string, format: date-time}
        lastSignInAt: {type: [string, 'null'], format: date-time}
...
    Session:
      type: object
      required: [id, userId, device, os, browser, ip, location, method, createdAt, lastActiveAt]
      properties:
        id: {type: string, examples: [ses_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
        userId: {type: string}
        device: {type: string}
        os: {type: string}
        browser: {type: string}
        ip: {type: string}
        location: {type: string}
        method: {$ref: '#/components/schemas/MethodKind'}
        createdAt: {type: string, format: date-time}
        lastActiveAt: {type: string, format: date-time}
        current: {type: boolean, description: Present and true for the session making the request.}
        client: {type: string, enum: [halo-cli], description: 'Present for sessions that `halo login` started through the device flow. Ending such a session, or all of a person''s sessions, also revokes its access and refresh tokens.'}
    SignInEvent:
      type: object
      required: [id, time, userId, email, appId, result, method, ip, location, device, risk]
      properties:
        id: {type: string, examples: [evt_01J9Z8Q4M2N5P7R9T1V3X5Z7B9]}
        time: {type: string, format: date-time}
        userId: {type: [string, 'null']}
        email: {type: string}
        appId: {type: [string, 'null'], description: Null for sign-ins to Halo itself.}
        result: {type: string, enum: [success, failure, interrupted]}
        method: {type: string}
        ip: {type: string}
        location: {type: string}
        device: {type: string, examples: [Chrome 130 · macOS]}
        risk: {type: string, enum: [none, low, medium, high]}
        reason: {type: string, description: Why the sign-in failed.}
...